bot.py

from telegram import Update
from telegram.ext import Updater, CommandHandler, CallbackContext
import requests

# /start command handle karne ke liye function
def start(update: Update, context: CallbackContext):
    update.message.reply_text('Welcome! Send me a TeraBox video link to download.')

# /download command handle karne ke liye function
def download_video(update: Update, context: CallbackContext):
    if context.args:
        video_link = context.args[0]
        # TeraBox se video download link ko fetch karne ki logic
        video_url = get_terabox_video_url(video_link)
        if video_url:
            update.message.reply_text('Downloading video...')
            context.bot.send_video(chat_id=update.message.chat_id, video=video_url)
        else:
            update.message.reply_text('Error: Could not retrieve video.')
    else:
        update.message.reply_text('Please provide a TeraBox video link.')

# TeraBox se video URL fetch karne ka example function
def get_terabox_video_url(link):
    # TeraBox API ya scraping technique ke through video URL lena hoga
    # Yahan par example ke liye direct link ko return kar rahe hain
    return link  # Isse replace karein apne logic ke saath

# Main function bot ko run karne ke liye
def main():
    token = "YOUR_BOT_API_TOKEN"  # Yahan par apna API token daalein
    updater = Updater(token, use_context=True)
    dispatcher = updater.dispatcher

    # Commands ke liye handlers add karna
    dispatcher.add_handler(CommandHandler('start', start))
    dispatcher.add_handler(CommandHandler('download', download_video))

    # Bot ko start karna
    updater.start_polling()
    updater.idle()

if __name__ == '__main__':
    main()
