requirements.txt

python-telegram-bot==20.0
requests==2.25.1
